pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sketch-photo-xai"
version = "0.1.0"
description = "Cross-domain forensic sketch-to-photo retrieval with Grad-CAM explanations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "matplotlib>=3.8",
  "numpy>=1.26",
  "pandas>=2.2",
  "Pillow>=10.0",
  "scikit-learn>=1.4",
  "torch>=2.2",
  "torchvision>=0.17",
  "tqdm>=4.66",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

